requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.4.0
google-generativeai==0.8.3
python-dotenv==1.0.1
lxml==5.1.0
